+++ app/Http/Controllers/IssueTypeController.php
@@ -10,6 +10,7 @@ use App\Http\Requests\UpdateIssueTypeRequest;
 use App\Models\IssueType;
 use App\Services\Repositories\Contracts\IIssueTypeRepository;
 use App\Services\Repositories\Contracts\IProjectRepository;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Log;
 use Illuminate\View\View;
@@ -30,9 +31,12 @@ class IssueTypeController extends Controller
      *
      * @param int $projectId
      * @return View
+     * @throws AuthorizationException
      */
     public function index(int $projectId): View
     {
+        $this->authorize(Action::VIEW_ANY, IssueType::class);
+
         return view('issueTypes.index', [
             'issueTypes' => $this->issueTypeRepository->getAllByProjectId($projectId)->get(),
             'project' => $this->projectRepository->find($projectId),
@@ -43,9 +47,12 @@ class IssueTypeController extends Controller
      * Show the form for creating a new resource.
      *
      * @return View
+     * @throws AuthorizationException
      */
     public function create(): View
     {
+        $this->authorize(Action::CREATE, IssueType::class);
+
         return view('issueTypes.create');
     }
 
@@ -55,9 +62,12 @@ class IssueTypeController extends Controller
      * @param StoreIssueTypeRequest $request
      * @param int $projectId
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function store(StoreIssueTypeRequest $request, int $projectId): RedirectResponse
     {
+        $this->authorize(Action::CREATE, IssueType::class);
+
         try {
             $this->issueTypeRepository->create($request->input());
 
@@ -81,9 +91,12 @@ class IssueTypeController extends Controller
      * @param int $projectId
      * @param IssueType $issueType
      * @return View
+     * @throws AuthorizationException
      */
     public function edit(int $projectId, IssueType $issueType): View
     {
+        $this->authorize(Action::UPDATE, $issueType);
+
         return view('issueTypes.edit', [
             'issueType' => $issueType,
         ]);
@@ -96,9 +109,12 @@ class IssueTypeController extends Controller
      * @param int $projectId
      * @param int $issueTypeId
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function update(UpdateIssueTypeRequest $request, int $projectId, int $issueTypeId): RedirectResponse
     {
+        $this->authorize(Action::UPDATE, $this->issueTypeRepository->find($issueTypeId));
+
         try {
             $this->issueTypeRepository->update($request->input(), $issueTypeId);
 
@@ -122,9 +138,12 @@ class IssueTypeController extends Controller
      * @param int $projectId
      * @param IssueType $issueType
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function destroy(int $projectId, IssueType $issueType): RedirectResponse
     {
+        $this->authorize(Action::DELETE, $issueType);
+
         try {
             $this->issueTypeRepository->delete($issueType);
